Better completion metadata
